Fix Site percentages type: live API returns fractional values
ListSitesAsync failed against a real controller with a JsonException because
statistics.percentages values (e.g. txRetry: 2.5) are decimals, not integers.
Percentages is now IReadOnlyDictionary<string, double>. Also stop emitting a
bare trailing '?' when all query parameters are empty.

+++ src/UniFi.Net.SiteManager/Models/Site.cs
@@ -47,7 +47,7 @@ public record SiteStatistics(
     SiteGateway Gateway,
     IReadOnlyList<object> InternetIssues,
     SiteIspInfo IspInfo,
-    IReadOnlyDictionary<string, int> Percentages
+    IReadOnlyDictionary<string, double> Percentages
 );
 
 /// <summary>
